refactor and move reusable parts
